+++ changelog.md
@@ -2,6 +2,8 @@
 
 * Add `HalfWord` and `ThirdWord` types,
   change types of `toZCurve`, `fromZCurve`, `toZCurve3`, `fromZCurve3` accordingly.
+* New module `Data.Chimera.Memoize` providing various
+  memoization combinators.
 
 # 0.3.4.0